Fixed connection error and data inconsistency

+++ sys_monitor/entities/cpu.py
@@ -1,6 +1,5 @@
 from psutil import cpu_times as cpu
 from .base_entity import BaseEntity
-from decimal import Decimal
 
 
 class CPU(BaseEntity):
@@ -9,4 +8,4 @@ class CPU(BaseEntity):
         return c.user + c.system + c.irq + c.softirq + c.nice
 
     def get_info(self, interval):
-        return round(Decimal(self.__get_busy_time() / interval), 4)
+        return round(self.__get_busy_time() / interval, 4)
